DB定義のタイポを修正
okンストラクタを作成してvalを使用できるように調整

+++ src/main/kotlin/com/example/kyudo_app/infrastructure/persistence/repository/AssignBelongingGroupRepository.kt
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository
 import java.util.*
 
 @Repository
-interface AssignBelongingGroupRepository : JpaRepository<BelongingGroupEntity?, Long?> {
-    override fun findById(groupId: Long?): Optional<BelongingGroupEntity?>
+interface AssignBelongingGroupRepository : JpaRepository<BelongingGroupEntity?, Int?> {
+    override fun findById(groupId: Int?): Optional<BelongingGroupEntity?>
 }
